refactor(core): move tests to highest level possible
