+++ harper-core/src/linting/noun_verb_confusion/mod.rs
@@ -17,10 +17,9 @@ pub(crate) const NOUN_VERB_PAIRS: &[(&str, &str)] = &[
     // Add more pairs here as needed
 ];
 
-pub use noun_instead_of_verb::NounInsteadOfVerb;
-pub use verb_instead_of_noun::VerbInsteadOfNoun;
+use noun_instead_of_verb::NounInsteadOfVerb;
+use verb_instead_of_noun::VerbInsteadOfNoun;
 
-// Merge the two linters with a descriptive message
 merge_linters! {
     NounVerbConfusion =>
         NounInsteadOfVerb,
@@ -30,14 +29,83 @@ merge_linters! {
 
 #[cfg(test)]
 mod tests {
-    use super::{NounInsteadOfVerb, effect_affect::EffectAffect};
+    use super::NounVerbConfusion;
     use crate::linting::tests::{assert_lint_count, assert_suggestion_result};
+
+    #[test]
+    fn corrects_good_advise() {
+        assert_suggestion_result("Good advise", NounVerbConfusion::default(), "Good advice");
+    }
+
+    #[test]
+    fn corrects_bad_advise() {
+        assert_suggestion_result(
+            "I just wanted to bring attention to this because it stood out to me as potentially bad advise.",
+            NounVerbConfusion::default(),
+            "I just wanted to bring attention to this because it stood out to me as potentially bad advice.",
+        );
+    }
+
+    #[test]
+    fn dont_flag_correct_better_advise() {
+        assert_lint_count(
+            "Hello! I am an engineer at Plexon and am conducting tests with Kilosort4 so we can better advise our clients.",
+            NounVerbConfusion::default(),
+            0,
+        );
+    }
+
+    #[test]
+    #[ignore = "'better advise' can be correct as above, or a mistake like here"]
+    fn correct_better_advise() {
+        assert_suggestion_result(
+            "Maybe this will be a decent idea, .or anybody has better advise :)",
+            NounVerbConfusion::default(),
+            "Maybe this will be a decent idea, .or anybody has better advice :)",
+        );
+    }
+
+    #[test]
+    fn dont_flag_correct_better_believe() {
+        assert_lint_count(
+            "You'd better believe this is bbedit-gist-maker.",
+            NounVerbConfusion::default(),
+            0,
+        );
+    }
+
+    #[test]
+    fn correct_strong_believe() {
+        assert_suggestion_result(
+            "cause my strong believe is that we must give any user to describe whether a post is meant factual",
+            NounVerbConfusion::default(),
+            "cause my strong belief is that we must give any user to describe whether a post is meant factual",
+        );
+    }
+
+    #[test]
+    fn correct_deep_breathe() {
+        assert_suggestion_result(
+            "Take deep breathe and Do it again!",
+            NounVerbConfusion::default(),
+            "Take deep breath and Do it again!",
+        );
+    }
+
+    #[test]
+    fn correct_bad_intend() {
+        assert_suggestion_result(
+            "What do you do if you only see slightly longer posts that may still be acceptable (and not bad intend from the poster)",
+            NounVerbConfusion::default(),
+            "What do you do if you only see slightly longer posts that may still be acceptable (and not bad intent from the poster)",
+        );
+    }
 
     #[test]
     fn corrects_belief_instead_of_verb() {
         assert_suggestion_result(
             "I belief in you.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "I believe in you.",
         );
     }
@@ -47,19 +115,19 @@ mod tests {
     fn corrects_breath_instead_of_verb() {
         assert_suggestion_result(
             "Remember to breath deeply.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Remember to breathe deeply.",
         );
     }
 
     #[test]
     fn does_not_flag_correct_believe() {
-        assert_lint_count("I believe in you.", NounInsteadOfVerb::default(), 0);
+        assert_lint_count("I believe in you.", NounVerbConfusion::default(), 0);
     }
 
     #[test]
     fn does_not_flag_correct_breath() {
-        assert_lint_count("Take a deep breath.", NounInsteadOfVerb::default(), 0);
+        assert_lint_count("Take a deep breath.", NounVerbConfusion::default(), 0);
     }
 
     // real-world example unit tests
@@ -68,7 +136,7 @@ mod tests {
     fn fix_when_i_breath_you_breath() {
         assert_suggestion_result(
             "When I breath, you breath!",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "When I breathe, you breathe!",
         );
     }
@@ -77,7 +145,7 @@ mod tests {
     fn fix_weather_climate_and_the_air_we_breath() {
         assert_suggestion_result(
             "Weather Climate and the Air We Breath",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Weather Climate and the Air We Breathe",
         );
     }
@@ -86,7 +154,7 @@ mod tests {
     fn fix_always_breath() {
         assert_suggestion_result(
             "breathing. remember to always breath.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "breathing. remember to always breathe.",
         );
     }
@@ -95,7 +163,7 @@ mod tests {
     fn fix_never_breath_a_word() {
         assert_suggestion_result(
             "And never breath a word about your loss; If you can force your heart and nerve and sinew.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "And never breathe a word about your loss; If you can force your heart and nerve and sinew.",
         );
     }
@@ -104,7 +172,7 @@ mod tests {
     fn fix_breath_for_seconds() {
         assert_suggestion_result(
             "Once turned on, the LED on the TX unit would breath for a few seconds, then go completely dead and not responding to objects in front of the sensors.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Once turned on, the LED on the TX unit would breathe for a few seconds, then go completely dead and not responding to objects in front of the sensors.",
         );
     }
@@ -113,7 +181,7 @@ mod tests {
     fn fix_breath_a_little_more_life() {
         assert_suggestion_result(
             "... up to 12% more performance, could breath a little more life into systems as old as Sandy Bridge.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "... up to 12% more performance, could breathe a little more life into systems as old as Sandy Bridge.",
         );
     }
@@ -122,7 +190,7 @@ mod tests {
     fn fix_the_diversity_we_breath() {
         assert_suggestion_result(
             "The Diversity We Breath: Community Diversity",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The Diversity We Breathe: Community Diversity",
         );
     }
@@ -131,7 +199,7 @@ mod tests {
     fn fix_belief() {
         assert_suggestion_result(
             "While I have no plans to return to aerospace I belief it gives me a unique perspective to many challenges.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "While I have no plans to return to aerospace I believe it gives me a unique perspective to many challenges.",
         );
     }
@@ -140,7 +208,7 @@ mod tests {
     fn fix_we_belief() {
         assert_suggestion_result(
             "In contrast to other vendors in e-mobility, we belief that true transparency is only trustworthy if the entire process ...",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "In contrast to other vendors in e-mobility, we believe that true transparency is only trustworthy if the entire process ...",
         );
     }
@@ -150,7 +218,7 @@ mod tests {
     fn fix_i_can_breath() {
         assert_suggestion_result(
             "Steps to reproduce Expected behaviour I can breath underwater.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Steps to reproduce Expected behaviour I can breathe underwater.",
         );
     }
@@ -159,7 +227,7 @@ mod tests {
     fn fix_caps_should_breath() {
         assert_suggestion_result(
             "CAPS 1 2 3 4 5 A B C D SHOULD BREATH A BIT MORE ?",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "CAPS 1 2 3 4 5 A B C D SHOULD BREATHE A BIT MORE ?",
         );
     }
@@ -168,7 +236,7 @@ mod tests {
     fn fix_can_you_advice_me() {
         assert_suggestion_result(
             "Can you advice me how to train?",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Can you advise me how to train?",
         );
     }
@@ -177,7 +245,7 @@ mod tests {
     fn fix_we_can_advice_you() {
         assert_suggestion_result(
             "Feel free to share more details about your use case, so we can advice you specifically based on your case.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Feel free to share more details about your use case, so we can advise you specifically based on your case.",
         );
     }
@@ -186,7 +254,7 @@ mod tests {
     fn fix_would_advice_against() {
         assert_suggestion_result(
             "So that I would advice against using a spindle in laser mode.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "So that I would advise against using a spindle in laser mode.",
         );
     }
@@ -195,7 +263,7 @@ mod tests {
     fn fix_advice_to_listen() {
         assert_suggestion_result(
             "The idea of this applicaton was inspired by Ray Dalio, who always advice to listen to people who know more than us by experience.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The idea of this applicaton was inspired by Ray Dalio, who always advise to listen to people who know more than us by experience.",
         );
     }
@@ -205,7 +273,7 @@ mod tests {
     fn dont_fix_advice_on_that() {
         assert_lint_count(
             "I don't do table returning functions in my code so can't offer you advice on that.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -214,7 +282,7 @@ mod tests {
     fn fix_advice_to_stick_with_openvscode() {
         assert_suggestion_result(
             "But unless you really need it, I would advice to stick with openvscode as there are nearly the same.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "But unless you really need it, I would advise to stick with openvscode as there are nearly the same.",
         );
     }
@@ -223,7 +291,7 @@ mod tests {
     fn fix_advice_to_back_up_os_image() {
         assert_suggestion_result(
             "I would advice to back up all OS image before any update, because you could lose something what was working previously.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "I would advise to back up all OS image before any update, because you could lose something what was working previously.",
         );
     }
@@ -232,7 +300,7 @@ mod tests {
     fn fix_advice_to_use_ms_store() {
         assert_suggestion_result(
             "I know we can always advice to use the MS store to download JASP instead",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "I know we can always advise to use the MS store to download JASP instead",
         );
     }
@@ -241,7 +309,7 @@ mod tests {
     fn fix_should_intent_be() {
         assert_suggestion_result(
             "Should intent be on the blocklist?",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Should intent be on the blocklist?",
         );
     }
@@ -250,7 +318,7 @@ mod tests {
     fn fix_if_you_intent() {
         assert_suggestion_result(
             "If you intent to use a 64 bits machine, change line 74",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "If you intend to use a 64 bits machine, change line 74",
         );
     }
@@ -259,7 +327,7 @@ mod tests {
     fn fix_what_you_would_intent_to_do() {
         assert_suggestion_result(
             "May I ask what you would intent to do with such a feature?",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "May I ask what you would intend to do with such a feature?",
         );
     }
@@ -268,7 +336,7 @@ mod tests {
     fn dont_flag_intent_records() {
         assert_lint_count(
             "there are always intent records associated to the txns",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -277,7 +345,7 @@ mod tests {
     fn fix_did_you_always_intent_to() {
         assert_suggestion_result(
             "Did you always intent to fight malware? No.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Did you always intend to fight malware? No.",
         );
     }
@@ -286,7 +354,7 @@ mod tests {
     fn fix_we_recommend_you_create_a_new_issue_on_github_explaining_what_you_intent_to_do() {
         assert_suggestion_result(
             "... we recommend you create a new issue on github explaining what you intent to do.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "... we recommend you create a new issue on github explaining what you intend to do.",
         );
     }
@@ -295,7 +363,7 @@ mod tests {
     fn fix_intent_to_use_non_imported_symbol() {
         assert_suggestion_result(
             "There's a warning reported for this code, saying that it may intent to use non-imported symbol",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "There's a warning reported for this code, saying that it may intend to use non-imported symbol",
         );
     }
@@ -306,7 +374,7 @@ mod tests {
     fn fix_to_emphasis_the() {
         assert_suggestion_result(
             "This one could be used in a dialog to emphasis the surprise.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "This one could be used in a dialog to emphasize the surprise.",
         );
     }
@@ -315,7 +383,7 @@ mod tests {
     fn allow_to_emphasis_at_end() {
         assert_lint_count(
             "Changes literal underscores to emphasis",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -324,7 +392,7 @@ mod tests {
     fn allow_to_intent_adjective() {
         assert_lint_count(
             "Cleanup passing statistics to intent aware iterator",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -333,7 +401,7 @@ mod tests {
     fn fix_to_advice_a_class() {
         assert_suggestion_result(
             "How to advice a class that have been intercepted by another javaagent",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "How to advise a class that have been intercepted by another javaagent",
         );
     }
@@ -342,7 +410,7 @@ mod tests {
     fn fix_to_breath_some() {
         assert_suggestion_result(
             "You go to the balcony to breath some fresh air and look down at the things outside.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "You go to the balcony to breathe some fresh air and look down at the things outside.",
         );
     }
@@ -351,7 +419,7 @@ mod tests {
     fn fix_to_emphasis_a() {
         assert_suggestion_result(
             "we'd like to emphasis a few points below",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "we'd like to emphasize a few points below",
         );
     }
@@ -360,7 +428,7 @@ mod tests {
     fn fix_to_advice_their() {
         assert_suggestion_result(
             "People who are managing this situation tend to advice their users to lock+unlock their screen",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "People who are managing this situation tend to advise their users to lock+unlock their screen",
         );
     }
@@ -371,7 +439,7 @@ mod tests {
     fn fix_positive_affect_on_small_businesses() {
         assert_suggestion_result(
             "The new law had a positive affect on small businesses.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The new law had a positive effect on small businesses.",
         );
     }
@@ -380,7 +448,7 @@ mod tests {
     fn fix_measured_the_affect_of_caffeine() {
         assert_suggestion_result(
             "We measured the affect of caffeine on reaction time.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "We measured the effect of caffeine on reaction time.",
         );
     }
@@ -389,7 +457,7 @@ mod tests {
     fn fix_side_affects_included_nausea() {
         assert_suggestion_result(
             "The side affects included nausea and fatigue.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The side effects included nausea and fatigue.",
         );
     }
@@ -398,7 +466,7 @@ mod tests {
     fn fix_cause_and_affect_not_same() {
         assert_suggestion_result(
             "Cause and affect are not the same thing.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Cause and effect are not the same thing.",
         );
     }
@@ -407,7 +475,7 @@ mod tests {
     fn fix_change_will_have_an_affect_on_revenue() {
         assert_suggestion_result(
             "The change will have an affect on our revenue.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The change will have an effect on our revenue.",
         );
     }
@@ -416,7 +484,7 @@ mod tests {
     fn fix_medicine_took_affect_within_minutes() {
         assert_suggestion_result(
             "The medicine took affect within minutes.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The medicine took effect within minutes.",
         );
     }
@@ -425,7 +493,7 @@ mod tests {
     fn fix_policy_will_come_into_affect() {
         assert_suggestion_result(
             "The policy will come into affect on October 1.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The policy will come into effect on October 1.",
         );
     }
@@ -434,7 +502,7 @@ mod tests {
     fn fix_rules_are_now_in_affect() {
         assert_suggestion_result(
             "The rules are now in affect.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The rules are now in effect.",
         );
     }
@@ -443,7 +511,7 @@ mod tests {
     fn fix_with_immediate_affect_office_closed() {
         assert_suggestion_result(
             "With immediate affect, the office is closed.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "With immediate effect, the office is closed.",
         );
     }
@@ -452,7 +520,7 @@ mod tests {
     fn fix_stunning_special_affects() {
         assert_suggestion_result(
             "The director used stunning special affects.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The director used stunning special effects.",
         );
     }
@@ -461,7 +529,7 @@ mod tests {
     fn fix_placebo_affect_can_be_powerful() {
         assert_suggestion_result(
             "The placebo affect can be powerful.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The placebo effect can be powerful.",
         );
     }
@@ -470,7 +538,7 @@ mod tests {
     fn fix_ripple_affect_across_market() {
         assert_suggestion_result(
             "We felt the ripple affect across the entire market.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "We felt the ripple effect across the entire market.",
         );
     }
@@ -479,7 +547,7 @@ mod tests {
     fn fix_snowball_affect_amplified_problem() {
         assert_suggestion_result(
             "The snowball affect amplified the problem.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The snowball effect amplified the problem.",
         );
     }
@@ -488,7 +556,7 @@ mod tests {
     fn fix_knock_on_affect_throughout_team() {
         assert_suggestion_result(
             "That decision had a knock-on affect throughout the team.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "That decision had a knock-on effect throughout the team.",
         );
     }
@@ -497,7 +565,7 @@ mod tests {
     fn fix_greenhouse_affect_warms_planet() {
         assert_suggestion_result(
             "The greenhouse affect warms the planet.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The greenhouse effect warms the planet.",
         );
     }
@@ -506,7 +574,7 @@ mod tests {
     fn fix_apology_had_little_affect() {
         assert_suggestion_result(
             "Her apology had little affect.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Her apology had little effect.",
         );
     }
@@ -515,7 +583,7 @@ mod tests {
     fn fix_settings_go_into_affect() {
         assert_suggestion_result(
             "The new settings go into affect after a restart.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The new settings go into effect after a restart.",
         );
     }
@@ -524,7 +592,7 @@ mod tests {
     fn fix_put_plan_into_affect() {
         assert_suggestion_result(
             "They put the new plan into affect last week.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "They put the new plan into effect last week.",
         );
     }
@@ -533,7 +601,7 @@ mod tests {
     fn fix_contract_comes_into_affect() {
         assert_suggestion_result(
             "The contract comes into affect at midnight.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The contract comes into effect at midnight.",
         );
     }
@@ -542,7 +610,7 @@ mod tests {
     fn fix_warning_had_no_affect_on_behavior() {
         assert_suggestion_result(
             "The warning had no affect on his behavior.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The warning had no effect on his behavior.",
         );
     }
@@ -551,7 +619,7 @@ mod tests {
     fn fix_inflation_had_opposite_affect() {
         assert_suggestion_result(
             "Inflation had the opposite affect than expected.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Inflation had the opposite effect than expected.",
         );
     }
@@ -560,7 +628,7 @@ mod tests {
     fn fix_regulation_remains_in_affect() {
         assert_suggestion_result(
             "The regulation remains in affect until further notice.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The regulation remains in effect until further notice.",
         );
     }
@@ -569,7 +637,7 @@ mod tests {
     fn fix_app_changes_take_affect() {
         assert_suggestion_result(
             "The app changes take affect next week.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "The app changes take effect next week.",
         );
     }
@@ -578,7 +646,7 @@ mod tests {
     fn fix_sound_affects_were_added() {
         assert_suggestion_result(
             "Sound affects were added in post.",
-            NounInsteadOfVerb::default(),
+            NounVerbConfusion::default(),
             "Sound effects were added in post.",
         );
     }
@@ -589,7 +657,7 @@ mod tests {
     fn corrects_noun_subject_effects_object() {
         assert_suggestion_result(
             "System outages effect our customers.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "System outages affect our customers.",
         );
     }
@@ -598,7 +666,7 @@ mod tests {
     fn corrects_effects_variant() {
         assert_suggestion_result(
             "This policy effects employee morale.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "This policy affects employee morale.",
         );
     }
@@ -607,7 +675,7 @@ mod tests {
     fn ignores_effect_change_idiom() {
         assert_lint_count(
             "Leaders work to effect change in their communities.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -616,7 +684,7 @@ mod tests {
     fn ignores_effect_noun_phrase() {
         assert_lint_count(
             "The effect your plan had was dramatic.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -625,7 +693,7 @@ mod tests {
     fn ignores_effect_as_result_noun() {
         assert_lint_count(
             "The effect was immediate and obvious.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -634,7 +702,7 @@ mod tests {
     fn ignores_to_effect_substitutions() {
         assert_lint_count(
             "or it may be desired to effect substitutions",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -643,7 +711,7 @@ mod tests {
     fn ignores_effect_followed_by_of_phrase() {
         assert_lint_count(
             "We measured the effect of caffeine on sleep.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -652,7 +720,7 @@ mod tests {
     fn ignores_side_effects_usage() {
         assert_lint_count(
             "Side effects may include mild nausea.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -661,7 +729,7 @@ mod tests {
     fn ignores_special_effects_phrase() {
         assert_lint_count(
             "She admired the special effects in the film.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -670,7 +738,7 @@ mod tests {
     fn ignores_effect_in_cause_and_effect() {
         assert_lint_count(
             "The diagram explains cause and effect relationships.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -679,7 +747,7 @@ mod tests {
     fn ignores_effects_with_pronoun_subject() {
         assert_lint_count(
             "Those effects were less severe than expected.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -688,7 +756,7 @@ mod tests {
     fn corrects_tariff_effect_import_prices() {
         assert_suggestion_result(
             "The new tariff will effect import prices next quarter.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The new tariff will affect import prices next quarter.",
         );
     }
@@ -697,7 +765,7 @@ mod tests {
     fn corrects_droughts_effect_crop_yields() {
         assert_suggestion_result(
             "Prolonged droughts severely effect crop yields across the valley.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Prolonged droughts severely affect crop yields across the valley.",
         );
     }
@@ -706,7 +774,7 @@ mod tests {
     fn corrects_caffeine_effect_sleep() {
         assert_suggestion_result(
             "Caffeine can effect your sleep architecture.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Caffeine can affect your sleep architecture.",
         );
     }
@@ -715,7 +783,7 @@ mod tests {
     fn corrects_bug_effect_devices() {
         assert_suggestion_result(
             "The firmware bug doesn't effect older devices.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The firmware bug doesn't affect older devices.",
         );
     }
@@ -724,7 +792,7 @@ mod tests {
     fn corrects_sarcasm_effect_morale() {
         assert_suggestion_result(
             "Her sarcasm seemed to effect the team's morale.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Her sarcasm seemed to affect the team's morale.",
         );
     }
@@ -733,7 +801,7 @@ mod tests {
     fn corrects_outage_effect_timeline() {
         assert_suggestion_result(
             "How will this outage effect our deployment timeline?",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "How will this outage affect our deployment timeline?",
         );
     }
@@ -742,7 +810,7 @@ mod tests {
     fn corrects_temperatures_effect_battery() {
         assert_suggestion_result(
             "Cold temperatures drastically effect lithium-ion battery performance.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Cold temperatures drastically affect lithium-ion battery performance.",
         );
     }
@@ -751,7 +819,7 @@ mod tests {
     fn corrects_policy_effect_eligibility() {
         assert_suggestion_result(
             "The policy change could effect your eligibility for benefits.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The policy change could affect your eligibility for benefits.",
         );
     }
@@ -760,7 +828,7 @@ mod tests {
     fn corrects_variables_effect_results() {
         assert_suggestion_result(
             "These confounding variables may effect the study's results.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "These confounding variables may affect the study's results.",
         );
     }
@@ -769,7 +837,7 @@ mod tests {
     fn corrects_fans_effect_concentration() {
         assert_suggestion_result(
             "The noisy HVAC fans constantly effect concentration in the lab.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The noisy HVAC fans constantly affect concentration in the lab.",
         );
     }
@@ -778,7 +846,7 @@ mod tests {
     fn corrects_hormones_effect_immunity() {
         assert_suggestion_result(
             "Stress hormones can effect immune response during recovery.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Stress hormones can affect immune response during recovery.",
         );
     }
@@ -787,7 +855,7 @@ mod tests {
     fn corrects_pacing_effect_engagement() {
         assert_suggestion_result(
             "The instructor's pacing tended to effect student engagement.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The instructor's pacing tended to affect student engagement.",
         );
     }
@@ -796,7 +864,7 @@ mod tests {
     fn corrects_humidity_effect_paint() {
         assert_suggestion_result(
             "Humidity levels directly effect paint curing time.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Humidity levels directly affect paint curing time.",
         );
     }
@@ -805,7 +873,7 @@ mod tests {
     fn corrects_exchange_effect_invoice() {
         assert_suggestion_result(
             "The exchange rate will surely effect the final invoice.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The exchange rate will surely affect the final invoice.",
         );
     }
@@ -814,7 +882,7 @@ mod tests {
     fn corrects_brightness_effect_contrast() {
         assert_suggestion_result(
             "Screen brightness settings can effect perceived contrast.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Screen brightness settings can affect perceived contrast.",
         );
     }
@@ -823,7 +891,7 @@ mod tests {
     fn corrects_medication_effect_him() {
         assert_suggestion_result(
             "The medication didn't effect him the way the doctor expected.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The medication didn't affect him the way the doctor expected.",
         );
     }
@@ -832,7 +900,7 @@ mod tests {
     fn corrects_payments_effect_credit() {
         assert_suggestion_result(
             "Late payments will negatively effect your credit score.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Late payments will negatively affect your credit score.",
         );
     }
@@ -841,7 +909,7 @@ mod tests {
     fn corrects_wording_effect_interpretation() {
         assert_suggestion_result(
             "Minor wording tweaks shouldn't effect the legal interpretation.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Minor wording tweaks shouldn't affect the legal interpretation.",
         );
     }
@@ -850,7 +918,7 @@ mod tests {
     fn corrects_traffic_effect_delivery() {
         assert_suggestion_result(
             "Traffic patterns often effect delivery windows downtown.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Traffic patterns often affect delivery windows downtown.",
         );
     }
@@ -859,7 +927,7 @@ mod tests {
     fn corrects_rumor_effect_confidence() {
         assert_suggestion_result(
             "The rumor started to effect investor confidence by noon.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The rumor started to affect investor confidence by noon.",
         );
     }
@@ -868,7 +936,7 @@ mod tests {
     fn corrects_allergies_effect_productivity() {
         assert_suggestion_result(
             "Seasonal allergies badly effect her productivity each April.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Seasonal allergies badly affect her productivity each April.",
         );
     }
@@ -877,7 +945,7 @@ mod tests {
     fn corrects_feedback_effect_roadmap() {
         assert_suggestion_result(
             "Your feedback won't immediately effect the roadmap.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Your feedback won't immediately affect the roadmap.",
         );
     }
@@ -886,7 +954,7 @@ mod tests {
     fn corrects_rules_effect_honeypot() {
         assert_suggestion_result(
             "I cant seem to get my additional rules to effect the honeypot",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "I cant seem to get my additional rules to affect the honeypot",
         );
     }
@@ -895,7 +963,7 @@ mod tests {
     fn corrects_bandwidth_effect_video() {
         assert_suggestion_result(
             "Fluctuating bandwidth can effect video call quality.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Fluctuating bandwidth can affect video call quality.",
         );
     }
@@ -904,7 +972,7 @@ mod tests {
     fn corrects_gradient_effect_sensor() {
         assert_suggestion_result(
             "The temperature gradient might effect the sensor's calibration.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The temperature gradient might affect the sensor's calibration.",
         );
     }
@@ -913,7 +981,7 @@ mod tests {
     fn corrects_delays_effect_satisfaction() {
         assert_suggestion_result(
             "Even tiny delays can effect user satisfaction metrics.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Even tiny delays can affect user satisfaction metrics.",
         );
     }
@@ -922,7 +990,7 @@ mod tests {
     fn corrects_architecture_effect_gps() {
         assert_suggestion_result(
             "The surrounding architecture can effect GPS accuracy.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The surrounding architecture can affect GPS accuracy.",
         );
     }
@@ -931,7 +999,7 @@ mod tests {
     fn corrects_lighting_effect_color() {
         assert_suggestion_result(
             "Lighting conditions strongly effect color perception.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Lighting conditions strongly affect color perception.",
         );
     }
@@ -940,7 +1008,7 @@ mod tests {
     fn corrects_coach_effect_roles() {
         assert_suggestion_result(
             "The new coach's strategy will effect players' roles.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The new coach's strategy will affect players' roles.",
         );
     }
@@ -949,7 +1017,7 @@ mod tests {
     fn corrects_overtraining_effect_reaction() {
         assert_suggestion_result(
             "Overtraining can effect reaction time and coordination.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Overtraining can affect reaction time and coordination.",
         );
     }
@@ -958,7 +1026,7 @@ mod tests {
     fn corrects_label_effect_behavior() {
         assert_suggestion_result(
             "The warning label may effect how consumers use the product.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The warning label may affect how consumers use the product.",
         );
     }
@@ -968,21 +1036,25 @@ mod tests {
     fn corrects_because_affect_is() {
         assert_suggestion_result(
             "I worry because affect is hidden.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "I worry because effect is hidden.",
         );
     }
 
     #[test]
     fn ignores_psychology_usage() {
-        assert_lint_count("The patient's affect is flat.", EffectAffect::default(), 0);
+        assert_lint_count(
+            "The patient's affect is flat.",
+            NounVerbConfusion::default(),
+            0,
+        );
     }
 
     #[test]
     fn corrects_positive_affect_on() {
         assert_suggestion_result(
             "The new law had a positive affect on small businesses.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The new law had a positive effect on small businesses.",
         );
     }
@@ -991,7 +1063,7 @@ mod tests {
     fn corrects_great_affect() {
         assert_suggestion_result(
             "badges that they provide to users to allow them to promote their projects to great affect",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "badges that they provide to users to allow them to promote their projects to great effect",
         );
     }
@@ -1000,7 +1072,7 @@ mod tests {
     fn corrects_affect_of() {
         assert_suggestion_result(
             "We measured the affect of caffeine on reaction time.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "We measured the effect of caffeine on reaction time.",
         );
     }
@@ -1009,7 +1081,7 @@ mod tests {
     fn corrects_side_affects() {
         assert_suggestion_result(
             "The side affects included nausea and fatigue.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The side effects included nausea and fatigue.",
         );
     }
@@ -1018,7 +1090,7 @@ mod tests {
     fn corrects_cause_and_affect() {
         assert_suggestion_result(
             "Cause and affect are not the same thing.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Cause and effect are not the same thing.",
         );
     }
@@ -1027,7 +1099,7 @@ mod tests {
     fn corrects_have_an_affect_on() {
         assert_suggestion_result(
             "The change will have an affect on our revenue.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The change will have an effect on our revenue.",
         );
     }
@@ -1036,7 +1108,7 @@ mod tests {
     fn corrects_took_affect() {
         assert_suggestion_result(
             "The medicine took affect within minutes.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The medicine took effect within minutes.",
         );
     }
@@ -1045,7 +1117,7 @@ mod tests {
     fn corrects_come_into_affect() {
         assert_suggestion_result(
             "The policy will come into affect on October 1.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The policy will come into effect on October 1.",
         );
     }
@@ -1054,7 +1126,7 @@ mod tests {
     fn corrects_in_affect_sentence() {
         assert_suggestion_result(
             "The rules are now in affect.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The rules are now in effect.",
         );
     }
@@ -1063,7 +1135,7 @@ mod tests {
     fn corrects_with_immediate_affect() {
         assert_suggestion_result(
             "With immediate affect, the office is closed.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "With immediate effect, the office is closed.",
         );
     }
@@ -1072,7 +1144,7 @@ mod tests {
     fn corrects_special_affects() {
         assert_suggestion_result(
             "The director used stunning special affects.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The director used stunning special effects.",
         );
     }
@@ -1081,7 +1153,7 @@ mod tests {
     fn corrects_placebo_affect() {
         assert_suggestion_result(
             "The placebo affect can be powerful.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The placebo effect can be powerful.",
         );
     }
@@ -1090,7 +1162,7 @@ mod tests {
     fn corrects_ripple_affect() {
         assert_suggestion_result(
             "We felt the ripple affect across the entire market.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "We felt the ripple effect across the entire market.",
         );
     }
@@ -1099,7 +1171,7 @@ mod tests {
     fn corrects_snowball_affect() {
         assert_suggestion_result(
             "The snowball affect amplified the problem.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The snowball effect amplified the problem.",
         );
     }
@@ -1108,7 +1180,7 @@ mod tests {
     fn corrects_knock_on_affect() {
         assert_suggestion_result(
             "That decision had a knock-on affect throughout the team.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "That decision had a knock-on effect throughout the team.",
         );
     }
@@ -1117,7 +1189,7 @@ mod tests {
     fn corrects_greenhouse_affect() {
         assert_suggestion_result(
             "The greenhouse affect warms the planet.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The greenhouse effect warms the planet.",
         );
     }
@@ -1126,7 +1198,7 @@ mod tests {
     fn corrects_little_affect() {
         assert_suggestion_result(
             "Her apology had little affect.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Her apology had little effect.",
         );
     }
@@ -1135,7 +1207,7 @@ mod tests {
     fn corrects_go_into_affect() {
         assert_suggestion_result(
             "The new settings go into affect after a restart.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The new settings go into effect after a restart.",
         );
     }
@@ -1144,7 +1216,7 @@ mod tests {
     fn corrects_put_plan_into_affect() {
         assert_suggestion_result(
             "They put the new plan into affect last week.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "They put the new plan into effect last week.",
         );
     }
@@ -1153,7 +1225,7 @@ mod tests {
     fn corrects_contract_into_affect() {
         assert_suggestion_result(
             "The contract comes into affect at midnight.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The contract comes into effect at midnight.",
         );
     }
@@ -1162,7 +1234,7 @@ mod tests {
     fn corrects_no_affect_on_behavior() {
         assert_suggestion_result(
             "The warning had no affect on his behavior.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The warning had no effect on his behavior.",
         );
     }
@@ -1171,7 +1243,7 @@ mod tests {
     fn corrects_opposite_affect() {
         assert_suggestion_result(
             "Inflation had the opposite affect than expected.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Inflation had the opposite effect than expected.",
         );
     }
@@ -1180,7 +1252,7 @@ mod tests {
     fn corrects_remains_in_affect() {
         assert_suggestion_result(
             "The regulation remains in affect until further notice.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The regulation remains in effect until further notice.",
         );
     }
@@ -1189,7 +1261,7 @@ mod tests {
     fn corrects_take_affect_next_week() {
         assert_suggestion_result(
             "The app changes take affect next week.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "The app changes take effect next week.",
         );
     }
@@ -1198,7 +1270,7 @@ mod tests {
     fn corrects_sound_affects() {
         assert_suggestion_result(
             "Sound affects were added in post.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Sound effects were added in post.",
         );
     }
@@ -1207,7 +1279,7 @@ mod tests {
     fn does_not_flag_best_affect() {
         assert_lint_count(
             "Using linear regression to predict and understand what factors best affect house price",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -1216,7 +1288,7 @@ mod tests {
     fn does_not_flag_sound_affect() {
         assert_lint_count(
             "The goal of this study was to learn what properties of sound affect human focus the most.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -1225,7 +1297,7 @@ mod tests {
     fn corrects_sound_affect() {
         assert_suggestion_result(
             "Diesel Generator's animation returns to 'idle' state, but it's sound affect remains in the 'work' state.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             "Diesel Generator's animation returns to 'idle' state, but it's sound effect remains in the 'work' state.",
         );
     }
@@ -1234,7 +1306,7 @@ mod tests {
     fn does_not_flag_affect_as_verb() {
         assert_lint_count(
             "The change will affect our revenue significantly.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -1243,7 +1315,7 @@ mod tests {
     fn does_not_flag_affects_as_verb() {
         assert_lint_count(
             "This policy directly affects remote workers.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -1252,7 +1324,7 @@ mod tests {
     fn does_not_flag_correct_effect_noun() {
         assert_lint_count(
             "The placebo effect can be powerful.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
@@ -1261,7 +1333,7 @@ mod tests {
     fn does_not_flag_sound_effects() {
         assert_lint_count(
             "Sound effects were added in post.",
-            EffectAffect::default(),
+            NounVerbConfusion::default(),
             0,
         );
     }
