+++ harper-core/src/linting/noun_verb_confusion/verb_instead_of_noun.rs
@@ -75,78 +75,3 @@ impl ExprLinter for VerbInsteadOfNoun {
         "Corrects verbs used instead of nouns when the two are related."
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::VerbInsteadOfNoun;
-    use crate::linting::tests::{assert_lint_count, assert_suggestion_result};
-
-    #[test]
-    fn corrects_good_advise() {
-        assert_suggestion_result("Good advise", VerbInsteadOfNoun::default(), "Good advice");
-    }
-
-    #[test]
-    fn corrects_bad_advise() {
-        assert_suggestion_result(
-            "I just wanted to bring attention to this because it stood out to me as potentially bad advise.",
-            VerbInsteadOfNoun::default(),
-            "I just wanted to bring attention to this because it stood out to me as potentially bad advice.",
-        );
-    }
-
-    #[test]
-    fn dont_flag_correct_better_advise() {
-        assert_lint_count(
-            "Hello! I am an engineer at Plexon and am conducting tests with Kilosort4 so we can better advise our clients.",
-            VerbInsteadOfNoun::default(),
-            0,
-        );
-    }
-
-    #[test]
-    #[ignore = "'better advise' can be correct as above, or a mistake like here"]
-    fn correct_better_advise() {
-        assert_suggestion_result(
-            "Maybe this will be a decent idea, .or anybody has better advise :)",
-            VerbInsteadOfNoun::default(),
-            "Maybe this will be a decent idea, .or anybody has better advice :)",
-        );
-    }
-
-    #[test]
-    fn dont_flag_correct_better_believe() {
-        assert_lint_count(
-            "You'd better believe this is bbedit-gist-maker.",
-            VerbInsteadOfNoun::default(),
-            0,
-        );
-    }
-
-    #[test]
-    fn correct_strong_believe() {
-        assert_suggestion_result(
-            "cause my strong believe is that we must give any user to describe whether a post is meant factual",
-            VerbInsteadOfNoun::default(),
-            "cause my strong belief is that we must give any user to describe whether a post is meant factual",
-        );
-    }
-
-    #[test]
-    fn correct_deep_breathe() {
-        assert_suggestion_result(
-            "Take deep breathe and Do it again!",
-            VerbInsteadOfNoun::default(),
-            "Take deep breath and Do it again!",
-        );
-    }
-
-    #[test]
-    fn correct_bad_intend() {
-        assert_suggestion_result(
-            "What do you do if you only see slightly longer posts that may still be acceptable (and not bad intend from the poster)",
-            VerbInsteadOfNoun::default(),
-            "What do you do if you only see slightly longer posts that may still be acceptable (and not bad intent from the poster)",
-        );
-    }
-}
